adding first shot at creating a module file!
The container.yaml spec is updated to include tags with hashes, and
we pull by the hash with singularity and then use lmod to install.
Since there are currently no other container technologies we default
to Singularity, and for the module system lmod, but this is created
in a modular (harhar) way so that we could add others in the future.
Each of the settings file and container.yaml file now have checking
with jsonschema, and there are classes for ContainerConfig and Tag/Tags.
I will next want to test this out on a cluster where I can actually
use Singularity

Signed-off-by: vsoch <vsoch@users.noreply.github.com>