tcl: Fix SINGULARITY_*OPTS env vars setup
Use the `setenv` modulefile command to define the SINGULARITY_OPTS and
SINGULARITY_COMMAND_OPTS environment variables in Tcl modulefiles.

It is important to get both of these environment variables defined even
if empty when the modulefile is loaded, as the various shell aliases
rely on these variables. On Csh shells, an *Undefined variable* error is
obtained if the environment variables are not defined.