tcl/lua: Add missing SINGULARITY_COMMAND_ARGS env. var
Add definition of the SINGULARITY_COMMAND_ARGS environment variable in
both Tcl and Lua modulefile templates. Set this variable to an empty
string like SINGULARITY_COMMAND_OPTS and SINGULARITY_OPTS. This is
important to set this variable when module is loaded as the -run shell
alias defined by this module relies on this variable. If not defined, an
*Undefined variable* error is obtained on Csh shells.