Adding first shot at new features (x11 and fake home) (#444)
* adding first shot at new features

this first commit will add a new feature, x11, which will be null by default. If the
value is True, we default to binding ~/.Xauthority. if it is found to be a string, we use that
instead. This means if there needs to be some custom bind like ~/.special-Xauthority TO the ~/.Xauthority,
both of those need to be represented in the single string.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* adding support for home feature

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* adding more docs about features to developer guide;

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* small tweaks to docs and adding home feature to biocontainers beast2

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

Co-authored-by: vsoch <vsoch@users.noreply.github.com>