first shot at adding wrapper scripts
this will support two kinds of wrapper scripts - global that live in the settings.yaml,
and container.yaml specific that go alongside that file. We basically will then generate
aliases for any executable names that were not created as wrappers. we will want to follow
this up with a command to list wrappers, and also include descriptions for the interested
user, and a flag to install to select on the fly

Signed-off-by: vsoch <vsoch@users.noreply.github.com>