proposal for handling default version on uninstall
instead of parsing each case, we can simply pretend that the last module installed
(determined by the creation or modify date) was just installed, and for all variants
of default_version this will work as it will either write this version as latest,
or just keep the dummy version. The exception is for first installed, for which we need
to pass a boolean that says "force calling this the first install." of course linux
does not do creation date well, so we do our best with modify date instead. Note I
have not tested this but am pushing the commit to further discuss!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>