Implemented "default_version==false" for TCL (#517)
* Implemented "default_version" for TCL
* refactor to use shared logic for default version template loading
I am also adding tests for both tcl and lmod default versions
* correct default version cases
* refactor to allow default_version to have multiple options
* change order of defaults for default_version
* proposal for handling default version on uninstall

instead of parsing each case, we can simply pretend that the last module installed
(determined by the creation or modify date) was just installed, and for all variants
of default_version this will work as it will either write this version as latest,
or just keep the dummy version. The exception is for first installed, for which we need
to pass a boolean that says "force calling this the first install." of course linux
does not do creation date well, so we do our best with modify date instead. Note I
have not tested this but am pushing the commit to further discuss!

* Clarified the logic in write_version_file and update_version_file
- New _set_default_version method to set the default version
- write_version_file and update_version_file now decide which is
  the first/last installed version (no more _default_version_first_installed)
* Expanded _cleanup to delete directories if they only contain ".version"
* TCL is now supported
* Merged update_version_file into write_version_file
* write_version_file can now detect the first/last version on its own
* Renamed the config file value to "module_sys" to match the other option
* Made can_be_deleted more generic, by taking the list of files to ignore as a parameter
* Reintroduced the tag parameter

Made it explicit that the tag write_version_file gets represents the version being installed

* Renamed _sys_module_default_version to _module_sys_default_version for consistency
* We can set default as None since it's not required.
* Use `latest_tag_installed` if possible

Co-authored-by: Matthieu Muffato <cortexspam-github@yahoo.fr>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>