Make "module-info shell" Tcl test compatible with Lmod<=8.7.55
Up to Lmod 8.7.55, the Tcl command "module-info shell" only returns the
current shell name and cannot test if a given string corresponds this
current shell.

A patch is now merged on Lmod, so newer version (8.7.56+) will cope with
the "module-info shell bash" syntax.

But in the mean time, I propose here to use a test Tcl syntax that will
be compatible with both Lmod and Environment Modules: "[module-info
shell] eq {bash}".

This change may only be of help for people willing to use the Tcl
modules produced by SHPC with the Lmod module tool.